backend/requirements.txt
python-dotenv==1.0.0
langchain==0.0.353
openai==1.6.1
chromadb==0.4.21
langchain-openai==0.0.3
